requirements.txt
ax>=0.52.0
ax_platform>=0.2.10
hydra-core>=1.3.1
omegaconf>=2.3.0
pandas==1.4.1
PyFoam>=2022.9
regex>=2022.10.31
dash>=2.7.0
dash_bootstrap_components>=1.4.2
